blogproject/requirements.txt
Django==2.2.3
django-haystack==3.0
django-pure-pagination==0.3.0
Faker==8.11.0
Markdown==3.3.4
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
